+++ src/EnterUtils.js
@@ -64,10 +64,11 @@ const utils = {
       });
     } else {
       a.map((m)=> {
-        if (!m || !m.key) {
+        if (!m || (!m.key && !m.props)) {
           return;
         }
-        for (var i = 0; i < b.length; i++) {
+
+        for (var i = b.length - 1; i >= 0; i--) {
           var cm = b[i];
           if (!cm || !cm.key || cm.key === m.key) {
             b.splice(i, 1);
@@ -79,59 +80,101 @@ const utils = {
       });
     }
   },
-  isPropsPushData: function (data, tagData, enterDataType) {
+  contrastEnterKeyArr: function (_a, _b, callback) {
+    var a = _a, b = _b.concat();
+    if (a.length === 0) {
+      b.map((m)=> {
+        callback.call(this, m);
+      });
+    } else {
+      a.map((m)=> {
+        if (!m || !m.props) {
+          return;
+        }
+        for (var i = b.length - 1; i >= 0; i--) {
+          var cm = b[i];
+          if (cm.props['enter-key'] === m.props['enter-key']) {
+            b.splice(i, 1);
+          }
+        }
+      });
+      b.map((m)=> {
+        callback.call(this, m);
+      });
+    }
+  },
+  isPropsPushData: function (data, enterDataType) {
     let _data = {};
-    if (data || tagData) {
-      data = data || {};
-      if (typeof data === 'boolean') {
-        data = {};
+    //if (data || tagData) {
+    //  data = data || {};
+    //  if (typeof data === 'boolean') {
+    //    data = {};
+    //  }
+    //  if (typeof tagData === 'string') {
+    //    tagData = JSON.parse(tagData);
+    //  }
+    //  tagData = tagData || {};
+    //  _data = utils.extend({}, [data, tagData]);
+    //  if (!_data.type && !_data.style) {
+    //    _data.type = enterDataType;
+    //  }
+    //}
+    if (data) {
+      //React.Children.forEach(data,(m)=> {
+      //  _data[m] = data[m];
+      //})
+      for (var a in data) {
+        _data[a] = data[a];
       }
-      if (typeof tagData === 'string') {
-        tagData = JSON.parse(tagData);
-      }
-      tagData = tagData || {};
-      _data = utils.extend({}, [data, tagData]);
-
       if (!_data.type && !_data.style) {
         _data.type = enterDataType;
       }
     }
     return _data;
   },
-  noPropsPushData: function (tagData, enterDataType) {
-    if (typeof tagData === 'string') {
-      tagData = JSON.parse(tagData);
+  //noPropsPushData: function (tagData, enterDataType) {
+  //  if (typeof tagData === 'string') {
+  //    tagData = JSON.parse(tagData);
+  //  }
+  //  if (tagData) {
+  //    if (typeof tagData === 'boolean') {
+  //      tagData = {};
+  //    }
+  //    if (!tagData.type && !tagData.style) {
+  //      tagData.type = enterDataType;
+  //    }
+  //  }
+  //  return tagData || {};
+  //},
+  leaveInherit: function (a, _root, b) {
+    if (a.type || b.type) {
+      a.type = a.type || _root.type || b.type;
     }
-    if (tagData) {
-      if (typeof tagData === 'boolean') {
-        tagData = {};
+    if (a.style || b.style) {
+      a.style = a.style || _root.style || b.style;
+    }
+    //console.log(typeof a.duration === 'number' || typeof b.duration === 'number')
+    if (typeof a.duration === 'number' || typeof b.duration === 'number') {
+      a.duration = typeof a.duration === 'number' ? a.duration : b.duration;
+    }
+    if (a.ease || b.ease) {
+      a.ease = a.ease || b.ease;
+    }
+    if (typeof a.delay === 'number' || b.delay) {
+      a.delay = typeof a.delay === 'number' ? a.delay : b.delay;
+    }
+    if (typeof a.queueId === 'number' || b.queueId) {
+      a.queueId = typeof a.queueId === 'number' ? a.queueId : b.queueId;
+    }
+    if (a.type || a.style) {
+      if (typeof a.duration !== 'number' && typeof _root.duration === 'number') {
+        a.duration = _root.duration;
       }
-      if (!tagData.type && !tagData.style) {
-        tagData.type = enterDataType;
+      if (!a.ease && _root.ease) {
+        a.ease = _root.ease;
       }
     }
-    return tagData || {};
-  },
-  leaveInherit: function (leave, enter) {
-    if (leave.type || enter.type) {
-      leave.type = leave.type || enter.type;
-    }
-    if (leave.style || enter.style) {
-      leave.style = leave.style || enter.style;
-    }
-    if (typeof leave.duration === 'number' || enter.duration) {
-      leave.duration = typeof leave.duration === 'number' ? leave.duration : enter.duration;
-    }
-    if (leave.ease || enter.ease) {
-      leave.ease = leave.ease || enter.ease;
-    }
-    if (typeof leave.delay === 'number' || enter.delay) {
-      leave.delay = typeof leave.delay === 'number' ? leave.delay : enter.delay;
-    }
-    if (typeof leave.queueId === 'number' || enter.queueId) {
-      leave.queueId = typeof leave.queueId === 'number' ? leave.queueId : enter.queueId;
-    }
-    return leave;
+    return a;
   }
 };
 export default utils;